requirements.txt
yt-dlp
tqdm
moviepy
spotipy
mutagen
Pillow